API reference

API reference

RandomTree.AbstractSimulator — Type.

Simulator on random/fixed trees

source

RandomTree.AbstractTree — Type.

Type of all trees represented by DFS degree sequences

source

RandomTree.BinaryTree — Type.

A conditional Galton-Watson tree with 2 Bernoulli(1/2) offspring distrbiution.

source

RandomTree.BinaryTree — Method.

BinaryTree(size::Int)

Construct a BinaryTree of size size.

source

RandomTree.CayleyTree — Type.

A conditional Galton-Watson tree with Poission(1) offspring distrbiution.

source

RandomTree.CayleyTree — Method.

CayleyTree(size::Int)

Construct a CayleyTree of size size.

source

RandomTree.CondGWTree — Type.

Type of conditional Galton-Watson trees represented by DFS degree sequences

source

RandomTree.DFSWalker — Type.

A DFSWalker does DFS walk on a tree to compute some property of the tree. It should implement at least one of the two functions.

visitfirst(walker::AbstractTreeWalker, degree_sequence, node_index, parent_index)
visitsecond(walker::AbstractTreeWalker, degree_sequence, node_index, parent_index)

The first is called when the DFS walk first enters a node. The second is call when the walk finally leaves the node.

source

RandomTree.DepthWalker — Type.

A walker computes the depths of all nodes in a tree.

source

RandomTree.FiniteTree — Type.

Type of fintie trees represented by DFS degree sequences

source

RandomTree.FirstOnlyDFSWalker — Type.

A DFS walker that only needed to be called when the walk enters a node.

source

RandomTree.FixedDirectedGraph — Type.

Representing a fixed directed graph with directed edges stored in an array

source

RandomTree.FixedGraph — Type.

Representing a fixed graph with edges stored in an array

source

RandomTree.FixedTreeGraph — Type.

Representing a fixed tree with directed edges stored in an array

source

RandomTree.FixedTreeGraph — Method.

FixedTreeGraph(size::Int, edges::Array{Tuple{Int, Int}})

Construct a FixedTreeGraph with edges edges of size size.

source

RandomTree.FixedTreeGraph — Method.

FixedTreeGraph(size::Int)

Construct a FixedTreeGraph with no edges of size.

source

RandomTree.FullDAryTree — Type.

A fixed full-d-ary tree, i.e., in a tree each node has either d or 0 children and each level is full.

source

RandomTree.FullDAryTree — Method.

FullDAryTree(height::Int, d::Int)

Construct a FullDAryTree of height height and maximum degree d.

source

RandomTree.GeneralCondGWTree — Type.

A conditional Galton-Watson tree.

source

RandomTree.GeneralCondGWTree — Method.

GeneralCondGWTree(size::Int, maxdegree::Int, dist::DiscreteUnivariateDistribution)

Construct a GeneralCondGWTree of size size, maximum degree maxdegree, and offspring distribution dist.

source

RandomTree.GeneralCondGWTree — Method.

GeneralCondGWTree(size::Int, maxdegree::Int, dist::DiscreteUnivariateDistribution)

Construct a GeneralCondGWTree of size size and offspring distribution dist.

source

RandomTree.GraphWalker — Type.

A walker that converts a tree represented in DFS degree sequence to a FixedTreeGraph.

source

RandomTree.HeightSimulator — Type.

Simulation of height of trees.

source

RandomTree.KcutSimulator — Type.

Simulation of the k-cut number of trees.

source

RandomTree.KcutWalker — Type.

A walker computes the (random) k-cut number of a tree.

source

RandomTree.KcutWalker — Method.

KcutWalker(k::Integer)

Construct a KcutWalker to compute the k-cut number of a tree.

source

RandomTree.LeafSimulator — Type.

Simulation of the number of leafs.

source

RandomTree.RandomRecursiveTree — Type.

A random recursive tree.

source

RandomTree.RandomRecursiveTree — Method.

RandomRecursiveTree(size::Int)

Construct a RandomRecursiveTree of size size.

source

RandomTree.SubtreeSizeWalker — Type.

A walker computes the size of all subtrees of a tree.

source

RandomTree.TotalPathSimulator — Type.

Simulation of the total path length of trees.

source

Base.size — Method.

size(t::FiniteTree)

Return the size of the tree t

source

Base.size — Method.

size(graph::FixedGraph)

Return the size of the graph graph

source

RandomTree.CatalanTree — Method.

CatalanTree(size::Int)

Construct a GeneralCondGWTree with Geometric(1/2) offspring distribution of size size.

source

RandomTree.DAryTree — Method.

DAryTree(size::Int, d::Int)

Construct a GeneralCondGWTree with Binomial(d, 1/d) offspring distribution of size size.

source

RandomTree.LogProductSimulator — Method.

LogProductSimulator(tree)

Construct a SubtreeSizeSimulator that simulate the sum of log(subtree size) of trees.

source

RandomTree.LogProductSimulator — Method.

LogProductSimulator(tree, power)

Construct a SubtreeSizeSimulator that simulate the sum of log(subtree size)^power of trees.

source

RandomTree.MotzkinTree — Method.

MotzkinTree(size::Int)

Construct a GeneralCondGWTree with DiscreteUniform(0, 2) offspring distribution of size size.

source

RandomTree.SizePowerSimulator — Method.

SizePowerSimulator(tree)

Construct a SubtreeSizeSimulator that simulate the sum of sqrt(subtree size) of trees.

source

RandomTree.SizePowerSimulator — Method.

SizePowerSimulator(tree, power)

Construct a SubtreeSizeSimulator that simulate the sum of (subtree size)^power of trees.

source

RandomTree.addedge! — Method.

addedge!(tree::FixedGraph, from, to)

Add an edge (from, to) in tree.

source

RandomTree.degrees — Function.

degrees(tree)

Return an iterator that traverses the degree sequence of tree in DFS order.

source

RandomTree.drawtree — Function.

Draw a grpahic representaion of a tree using graphivz and its Python Interface

source

RandomTree.drawtree — Function.

drawtree(tree_digraph::FixedGraph, show_label=false)

Draw a graphic representation of tree_digraph

source

RandomTree.drawtree — Function.

drawtree(tree::FiniteTree, show_label=false)

Draw a graphic representation of the random/fixed tree tree.

source

RandomTree.drawtree — Function.

drawtree(deg_seq::Vector{Int}, show_label=false)

Draw a graphic representation of tree with DFS degree sequence deg_seq

source

RandomTree.edges — Method.

Return the array of edges of the graph

source

RandomTree.main — Method.

Entry point of command line script

source

RandomTree.maxdegree — Method.

maxdegree(t::FiniteTree)

Return the maximum degree of the tree t

source

RandomTree.name — Method.

name(t::FiniteTree)

Return the name of the tree t

source

RandomTree.nodes — Method.

Return the nodes of the graph as an iterator

source

RandomTree.positions — Method.

positions(tree::FixedTreeGraph)::Array{Int, 2}

First sort the edges in the tree and then return an array containing the position and length of the segment in the edge array corresponding to each node.

source

RandomTree.print_simulation — Method.

print_simulation(sim::AbstractSimulator, nsample)

Print nsample samples in simulation sim.

source

RandomTree.result — Method.

result(walker::AbstractTreeWalker)

Return of the result of the walk on the tree.

source

RandomTree.simulation — Method.

simulation(sim::AbstractSimulator, nsample)

Return nsample samples in simulation sim.

source

RandomTree.simulation — Method.

simulation(sim::SubtreeSizeSimulator)

Return one sample in simulation sim.

source

RandomTree.sortedge! — Method.

sortedge!(tree::FixedTreeGraph)

Sort the edges in tree according to starting nodes.

source

RandomTree.treegraph — Function.

treegraph(tree)::FixedTreeGraph

Either convert a fixed tree to a FixedTreeGraph, or generate a random FixedTreeGraph according to tree.

source

RandomTree.walk — Method.

walk(tree::FiniteTree, walker::DFSWalker)

Make walker do a DFS walk on tree.

source

RandomTree.FiniteFixedTree — Type.

Type of fintie fixed trees represented by DFS degree sequences

source

RandomTree.FiniteRandomTree — Type.

Type of fintie random trees represented by DFS degree sequences

source

RandomTree.SubtreeSizeSimulator — Type.

Simulation of the additive function in the form of sum of func(subtree size) applied to random trees.

source

RandomTree.TreeSpec — Type.

The specification of a tree.

source

RandomTree.TreeSpec — Method.

TreeSpec(size::Int, name::String)

Construct a TreeSpec of size size and name name.

source

RandomTree.parse_commandline — Method.

Parse command line arguments

source

RandomTree.visitfirst — Method.

visitfirst(walker::AbstractTreeWalker, degree_sequence, node_index, parent_index)

The function is called when the walker first enters a node.

source

RandomTree.visitsecond — Method.

visitsecond(walker::AbstractTreeWalker, degree_sequence, node_index, parent_index)

The function is called when the walker finally leaves a node.

source